backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
httpx==0.25.2
openpyxl==3.1.5
slowapi==0.1.9